requirements.txt
blinker
certifi
charset-normalizer
click
colorama
dnspython
Flask
idna
itsdangerous
Jinja2
MarkupSafe
pymongo
python-dotenv
requests
urllib3
Werkzeug
